fix(báo cáo tài chính): ghi nhận THU/CHI dồn tích theo ngày
Báo cáo trộn hai gốc kế toán: THU tính một cục tại ngày mời/gia hạn còn CHI
tính theo ngày phát hành hoá đơn Stripe. Hai hoá đơn ChatGPT đang hiệu lực đều
phát hành trong tháng 7 (GPT1 11/7 phủ tới 11/8, CHATGPT PRO 25/7 phủ tới 25/8)
nên khoảng 01→11/08 ra CHI = 0 và biên lợi nhuận 100% ảo.

Nay cả hai vế rải đều theo ngày, chỉ phần rơi trong [from, to] được tính:
phí kỳ member rải trên [start_at, end_at), hoá đơn ChatGPT rải trên
[period_start, period_end). Hoá đơn cũ chưa có period_* vẫn tính gọn 1 ngày.
seat_months thành seat·ngày ÷ 30 (thập phân) nên giá thu TB/seat và giá vốn
TB/seat cùng mẫu số.

Số thật 01→11/08 sau khi sửa: thu 21.630.868, chi 22.758.837, lỗ 1.127.969
(−5,2%) thay vì lãi 14.010.000 (+100%).

+++ apps/api/app/schemas.py
@@ -1315,7 +1315,8 @@ class FinancialReportOut(BaseModel):
     by_agent: list[FinancialReportAgent]
     # Số workspace chưa đồng bộ hoá đơn 'paid' → giá vốn thiếu (CHI có thể thấp hơn thực).
     cost_missing_workspaces: int
-    # Tổng "seat-tháng" có phát sinh THU trong kỳ (Σ số tháng của mọi kỳ được tính).
-    seat_months: int = 0
+    # Seat-tháng DỒN TÍCH có phát sinh THU trong kỳ = Σ seat-ngày phủ [from, to] ÷ 30.
+    # Thập phân (khoảng lẻ ngày → không tròn tháng).
+    seat_months: float = 0
     # Giá vốn TB mỗi seat/tháng = cost ÷ seat_months (None khi seat_months = 0).
     avg_cost_per_seat: int | None = None
